Fixed: parallel::distributed::refine_and_coarsen_fixed_fraction() contained a rounding bug that often produced wrong results.
git-svn-id: https://svn.dealii.org/trunk@24920 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/source/distributed/grid_refinement.cc
@@ -710,9 +710,8 @@ namespace parallel
 	      RefineAndCoarsenFixedFraction::
 	      master_compute_threshold (locally_owned_indicators,
 					global_min_and_max,
-					static_cast<unsigned int>
-					(top_fraction_of_error *
-					 total_error),
+					top_fraction_of_error *
+					 total_error,
 					mpi_communicator);
 
 					     // compute bottom
@@ -727,9 +726,8 @@ namespace parallel
 		RefineAndCoarsenFixedFraction::
 		master_compute_threshold (locally_owned_indicators,
 					  global_min_and_max,
-					  static_cast<unsigned int>
-					  ((1-bottom_fraction_of_error) *
-					   total_error),
+					  (1-bottom_fraction_of_error) *
+					   total_error,
 					  mpi_communicator);
 	    else
 	      {
